+++ salsa20/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.6.0 (2020-08-14)
+## 0.6.0 (2020-08-25)
 ### Changed
-- Bump `stream-cipher` dependency to v0.6 ([#161])
+- Bump `stream-cipher` dependency to v0.7 ([#161], [#164])
 
 [#161]: https://github.com/RustCrypto/stream-ciphers/pull/161
+[#164]: https://github.com/RustCrypto/stream-ciphers/pull/164
 
 ## 0.5.2 (2020-06-11)
 ### Changed